+++ src/main/java/com/javainiaisuzspringom/tripperis/config/SecurityConfig.java
@@ -47,6 +47,9 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.POST, "/api/status-code/**").hasRole("ADMIN")
                 .antMatchers("/api/**").hasAnyRole("USER", "ADMIN")
                 .and()
+                .logout()
+                .logoutSuccessUrl("/auth/login")
+                .and()
                 .apply(new JwtConfigurer(jwtTokenProvider));
     }
